requirements.txt
biopython>=1.81
importlib-metadata; python_version<"3.8"
pyhmmer>=0.10.4
